+++ src/core/linear_solvers/KLUSolver.hpp
@@ -65,8 +65,9 @@ class LS2G_API KLULinearSolver final
 
         // public api
         ErrorType reset();
-        ErrorType initialize(const Eigen::SparseMatrix<real_type>& J);
-        ErrorType refactor(const Eigen::SparseMatrix<real_type>& J);
+        ErrorType analyze(const Eigen::SparseMatrix<real_type>& J);   // reordering + symbolic factorization (structure only)
+        ErrorType factorize(const Eigen::SparseMatrix<real_type>& J); // numeric factorization (requires values)
+        ErrorType refactorize(const Eigen::SparseMatrix<real_type>& J);  // re-numeric factorization, reuses symbolic
         ErrorType solve(RealVect & b);
 
         // can this linear solver solve problem where RHS is a matrix